Performs hard reboot if libvirt soft reboot raises libvirtError
Currently the libvirt driver will only attempt a hard reboot if a soft
reboot fails and returns False. If a libvirt exception is raised,
however, the hard reboot is not attempted. This patch wraps the soft
reboot in a try block to continue with the hard reboot attempt if a
libvirtError is raised.

Fixes bug 1202974

(cherry picked from commit 9925c9f1b353240ebb1eccde50574055ce271313)

Conflicts:
	nova/tests/test_libvirt.py

Change-Id: Ifaff75b33b5123cdd3293e8f263ccb4f72ae726a