Changes attach_volume method signature to match icehouse version.
This allows icehouse api cells to talk to havana compute cells.
The the extra arguments are optional and don't break havana if not
specified. The actual change for this is quite large and may not be
suitable for backporting.
See: 61e6f5b1e1a0a43b837e7c428a9624af06bbb4cf

Change-Id: I46778281c8186db457bdf0562e19d25adecd3029
Closes-bug: #1308846