VMware: fix exception when multiple compute nodes are running
A number of operations in the VMwareVCDriver class first validate
that instances node is the cluster that is mapped to the compute
node. This is problematic when the compute nodes have different
configurations, for example, each compute node is mapped to a
different cluster.

In this case many operations that are just performing instance operations
will fail. This patch ensure that all instance operations that do not
require a cluster or volume will make use of the base _vmops class.
This is due to the fact that it only requires the instance details
to interface with the VC and there are no specific cluster operations.

Change-Id: I2bc38a480f2feb12ea41e7d28f80b29dd49a79b8
Closes-bug: #1345460
(cherry picked from commit 8e4a9156f4dccf003970848c28b8a9d15c55212d)