Fix detach_sriov_ports to get context to be able to get image metadata
The previous implementation took the context using
get_admin_context() which return context with admin flag set,
but all the other attributes are None.
This is not sufficient to get image metadata,
it is need context with elevated permission.
The change here is to update API to pass the context to suspend so
it will be able to pass it to detach_sriov_ports.
Also update the snapshot method to pass context to detach_sriov_ports

(cherry picked from commit 6f002d26f28998c99d9922fdd49b0805c44ff22f)

Closes-Bug:#1406486
Change-Id: I4757a7646d7bee66db03c5d2410de7378c039d41