Hyper-V: Fix missing WMI namespace issue on Windows 2008 R2
The Hyper-V driver uses the Microsoft\Windows\SMB WMI namespace
in order to handle SMB shares. The issue is that this namespace is
not available on Windows versions prior to Windows Server 2012.

For this reason, the Hyper-V driver fails to initialize on Windows
Server 2008 R2.

This patch fixes the issue by properly handling the PathUtils
initialization.

Closes-Bug: #1463044
Change-Id: Ia99576589af6049ee07337c631ed7d5d6cf602d9
(cherry picked from commit 0625df995629291a6112197f61099cd82b31679b)