Kilo-Removing extension "OS-EXT-VIF-NET" from v2.1 ext list
Extension "OS-EXT-VIF-NET" is not ported to v2.1, due to that
there is difference between v2 and v2.1 as mentioned in bug#1470690.

But v2.1 extension-list has "OS-EXT-VIF-NET" extension which convey that
this extension is also loaded for v2.1 and user will get the expected response
according to "OS-EXT-VIF-NET" extension which is not true as this extension
is not actually ported to v2.1.

It is decided that fix for that difference should go in microversion not
in base v2.1 as it is already released.

This patch remove extension "OS-EXT-VIF-NET" from v2.1 extension-list which
will actually convey that this extension is missing in v2.1.

This is back-porting Ic99ac1179d02d907422911fe1369b64479fd5f33 in Kilo stable
branch.

Partial-Bug: 1470690

Change-Id: Ic590bb5df2b2df917af9ce9941da6afba4e382a5